fix/admin (#31) : Suppression de ProfileDashboardController.
