+++ src/Controller/Admin/DashboardController.php
@@ -2,34 +2,37 @@
 
 namespace App\Controller\Admin;
 
-use EasyCorp\Bundle\EasyAdminBundle\Config\Dashboard;
-use EasyCorp\Bundle\EasyAdminBundle\Config\MenuItem;
-use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractDashboardController;
+use App\Entity\User;
+use App\Entity\Kanji;
+use App\Entity\KanjiKey;
+use App\Entity\Ideogramme;
 use Symfony\Component\HttpFoundation\Response;
+use App\Controller\Admin\ProfileCrudController;
 use Symfony\Component\Routing\Annotation\Route;
+use EasyCorp\Bundle\EasyAdminBundle\Config\Crud;
+use EasyCorp\Bundle\EasyAdminBundle\Config\Action;
+use EasyCorp\Bundle\EasyAdminBundle\Config\MenuItem;
+use EasyCorp\Bundle\EasyAdminBundle\Config\Dashboard;
+use EasyCorp\Bundle\EasyAdminBundle\Router\AdminUrlGenerator;
+use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractDashboardController;
 
 class DashboardController extends AbstractDashboardController
 {
     #[Route('/admin', name: 'admin')]
     public function index(): Response
     {
-        return parent::index();
+        /** @var User $user  */
+        $user = $this->getUser();
 
-        // Option 1. You can make your dashboard redirect to some common page of your backend
-        //
-        // $adminUrlGenerator = $this->container->get(AdminUrlGenerator::class);
-        // return $this->redirect($adminUrlGenerator->setController(OneOfYourCrudController::class)->generateUrl());
-
-        // Option 2. You can make your dashboard redirect to different pages depending on the user
-        //
-        // if ('jane' === $this->getUser()->getUsername()) {
-        //     return $this->redirect('...');
-        // }
-
-        // Option 3. You can render some custom template to display a proper dashboard with widgets, etc.
-        // (tip: it's easier if your template extends from @EasyAdmin/page/content.html.twig)
-        //
-        // return $this->render('some/path/my-dashboard.html.twig');
+        /** @var AdminUrlGenerator $adminUrlGenerator */
+        $adminUrlGenerator = $this->container->get(AdminUrlGenerator::class);
+        return $this->redirect(
+            $adminUrlGenerator
+                ->setController(ProfileCrudController::class)
+                ->setEntityId($user->getId())
+                ->setAction(Action::DETAIL)
+                ->generateUrl()
+        );
     }
 
     public function configureDashboard(): Dashboard
@@ -40,7 +43,27 @@ class DashboardController extends AbstractDashboardController
 
     public function configureMenuItems(): iterable
     {
-        yield MenuItem::linkToDashboard('Dashboard', 'fa fa-home');
-        // yield MenuItem::linkToCrud('The Label', 'fas fa-list', EntityClass::class);
+        return [
+            MenuItem::linkToDashboard('Dashboard', 'fa fa-home'),
+
+            MenuItem::section('Idéogrammes'),
+            MenuItem::linkToCrud('La liste complète', 'fa fa-file-text', Ideogramme::class),
+
+            MenuItem::section('Les kanji'),
+            MenuItem::linkToCrud('Voir la liste', 'fa fa-file-text', Kanji::class),
+            MenuItem::linkToCrud('Ajouter', 'fas fa-plus', Kanji::class)->setAction(ACTION::NEW),
+
+            MenuItem::section('Les clés'),
+            MenuItem::linkToCrud('Voir la liste', 'fa fa-file-text', KanjiKey::class),
+            MenuItem::linkToCrud('Ajouter', 'fas fa-plus', KanjiKey::class)->setAction(ACTION::NEW),
+        ];
+    }
+
+    public function configureCrud(): Crud
+    {
+        return Crud::new()
+            ->setPaginatorPageSize(15)
+            ->setEntityPermission('ROLE_ADMIN')
+        ;
     }
 }
